support for oblivion remastered
This enables XGP saves of oblivion remastered to be output so that you can use it in other directories / releases.

It copies the forza handler, but instead of .data it puts it into .sav

It updates the games.json accordingly.

tested and was able to play using the converted files.